requirements.txt
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
gradio>=4.0.0
langchain-core>=0.1.0
